fix: couple swapper addresses to the chain ID
`<T>Swapper` constructors now require the current chain ID as their second parameter, which they `assert()` as correct because the `<T>SwapperDeployer` always sends the correct ID. Despite the redundant behaviour, the swapper address is now coupled to the particular chain because the ID is used in predicting the CREATE2 address.

Note: this still requires testing of the unhappy path.